Rename `libddprof_helpers.h` to `libdatadog_helpers.h`
One more step towards the rename of libddprof to libdatadog.
