+++ ext/ddtrace_profiling_native_extension/libdatadog_helpers.h
